CE - Evals/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ce-evals"
version = "0.1.0"
description = "LLM-as-a-Judge evaluation toolkit for multi-agent protocols"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.39.0",
    "openai>=1.0",
    "google-genai>=1.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio"]

[tool.setuptools.packages.find]
where = ["src"]
